调整接口调用返回码数据类型及编码位数

+++ driving-api/src/main/java/com/udriving/drivingapi/controller/response/Response.java
@@ -10,7 +10,7 @@ public class Response implements ResponseConstant {
     /**
      * 接口响应状态
      */
-    byte code = SUCCEED;
+    short code = SUCCEED;
     /**
      *
      */
